refactored for astropath
