--- frb/associate/bayesian.py
@@ -1,323 +1,0 @@
-"""Methods related to Bayesian association analysis"""
-
-import numpy as np
-
-import copy
-
-from astropy import units
-from astropy.coordinates import SkyCoord
-
-from IPython import embed
-
-
-def add_contam(ncontam, frb_coord, cand_gal, cand_coord, fov,
-               rstate=None, rmag=25.):
-    if ncontam != 1:
-        import pdb; pdb.set_trace()  # Not ready for more than 1
-    if rstate is None:
-        rstate = np.random.RandomState()
-    # Generate a random contaminant -- Square inside a circle to keep it simple
-    s = 2*fov / np.sqrt(2)
-    x,y  = rstate.uniform(size=2*ncontam, low=0, high=s)
-    r = np.sqrt(x**2 + y**2)
-    pa = float(rstate.uniform(size=ncontam, low=0., high=360.))
-    #
-    c_coord = frb_coord.directional_offset_by(pa*units.deg, r*units.arcsec)
-    # Add to table
-    row = copy.deepcopy(cand_gal[0])
-    row['DES_r'] = rmag
-    row['ra'] = c_coord.ra.value
-    row['dec'] = c_coord.dec.value
-    cand_gal.add_row(row)
-    # Coordinate
-    cand_coord = SkyCoord([cand_coord, c_coord])
-    # Return
-    return cand_gal, cand_coord
-
-
-
-def raw_prior_Oi(Pchance, Sigma_m, method):
-    """
-    Raw prior for a given set of Pchance values
-
-    Proper normalization requires P(U) so that is done below
-
-    Args:
-        Pchance (np.ndarray):
-            Chance probability
-        Sigma_m (np.ndarray):
-            Number density of sources on the sky brighter than m
-        method (str):
-            linear
-            orig_inverse :: Assign inverse to P_chance
-            inverse :: Assign inverse to Sigma_m
-            identical :: All the same
-
-    Returns:
-        np.ndarray:
-
-    """
-    if method == 'linear':
-        return 1 - Pchance
-    elif method == 'orig_inverse':
-        return 1. / Pchance
-    elif method == 'inverse':
-        return 1. / Sigma_m
-    elif method == 'identical':
-        return np.ones_like(Pchance)
-    else:
-        raise IOError("Bad method {} for prior_Oi".format(method))
-
-
-'''
-def prior_S_n(rlim, theta_max, rmax=30.):
-    rval = np.linspace(rlim, rmax, 100)
-    # Nbar
-    sigma = bloom_sigma(rval)
-    Nbar = np.pi * theta_max ** 2 * sigma
-    # Average
-    P_S = np.sum(sigma * np.exp(-Nbar)) / np.sum(sigma)
-    # Return
-    return P_S
-'''
-
-
-# TODO -- JIT this
-def pw_Oi(r_w, theta_half, theta_prior, scale_half=1.):
-    """
-    Calculate p(w|O_i) for a given galaxy
-
-    Args:
-        r_w (np.ndarray):
-            offset from galaxy center in arcsec
-        theta_half (float):
-            Half-light radius of this galaxy in arcsec
-        theta_prior (dict):
-            Parameters for theta prior
-        scale_half (float):
-
-    Returns:
-        np.ndarray: Probability values; un-normalized
-
-    """
-    p = np.zeros_like(r_w)
-    ok_w = r_w < theta_prior['max']*theta_half
-    norm = 0
-    if theta_prior['method'] == 'core':
-        # Wolfram
-        norm = theta_half * np.log(theta_prior['max']+1)
-        if np.any(ok_w):
-            p[ok_w] = theta_half / (r_w[ok_w] + theta_half) / norm
-    elif theta_prior['method'] == 'uniform':
-        norm = theta_half * theta_prior['max']
-        if np.any(ok_w):
-            p[ok_w] = 1. / norm
-    elif theta_prior['method'] == 'exp':
-        #norm = theta_half  - theta_half * (theta_prior['max']+1) * np.exp(-theta_prior['max'])
-        norm = theta_half*scale_half * (scale_half - (
-                scale_half+theta_prior['max'])*np.exp(-theta_prior['max']/scale_half))
-        if np.any(ok_w):
-            p[ok_w] = (r_w[ok_w] / theta_half) * np.exp(-r_w[ok_w]/(scale_half*theta_half)) / norm
-    else:
-        raise IOError("Bad theta method")
-    #
-    if norm == 0:
-        raise ValueError("You forgot to normalize!")
-    # Return
-    return p
-
-
-def px_Oi(box_radius, frb_coord, eellipse, cand_coords,
-          theta_prior, step_size=0.1, return_grids=False):
-    """
-    Calculate p(x|O_i), the primary piece of the analysis
-
-    Main concept:
-        1. Set an area to analyze
-        2. Discretize it to 0.1"
-        3. Convolve
-
-    Args:
-        box_radius (float):
-            Maximum radius for analysis, in arcsec
-        eellipse (dict):
-            Error ellipse for the FRB
-            a, b, theta
-        frb_coord (SkyCoord):
-        cand_coords (SkyCoord):
-            Coordinates of the candidate hosts
-        theta_prior (dict):
-            Parameters for theta prior
-        step_size (float, optional):
-            Step size for grid, in arcsec
-
-    Returns:
-        np.ndarray:
-
-    """
-    # Error ellipse
-    pa_ee = eellipse['theta'] # deg
-    dtheta = 90. - pa_ee  # Place a of ellipse along the x-axis
-    # Set Equinox (for spherical offsets)
-    frb_coord.equinox = cand_coords[0].equinox
-    #
-    ngrid = int(np.round(2*box_radius / step_size))
-    x = np.linspace(-box_radius, box_radius, ngrid)
-    xcoord, ycoord = np.meshgrid(x,x)
-
-    # Build the grid around the FRB (orient semi-major axis "a" on our x axis)
-    l_w = np.exp(-xcoord ** 2 / (2 * eellipse['a'] ** 2)) * np.exp(-ycoord ** 2 / (2 * eellipse['b'] ** 2))
-
-    p_xMis, grids = [], []
-    # TODO -- multiprocess this
-    for icand, cand_coord in enumerate(cand_coords):
-
-        # Calculate observed FRB location
-        #dra, ddec = cand_coord.spherical_offsets_to(frb_coord)
-        #xFRB = -dra.to('arcsec').value
-        #yFRB = ddec.to('arcsec').value
-
-        # #####################
-        # l(w) -- 2D Gaussian
-
-        # Rotate the galaxy
-        r = frb_coord.separation(cand_coord).to('arcsec')
-        pa_gal = frb_coord.position_angle(cand_coord).to('deg')
-        new_pa_gal = pa_gal + dtheta * units.deg
-
-        #r_wsq = (xcoord-xFRB)**2 + (ycoord-yFRB)**2
-        #l_w = np.exp(-r_wsq/(2*sigR**2)) / sigR / np.sqrt(2*np.pi)
-
-        # p(w|M_i)
-        # x, y gal
-        x_gal = -r.value * np.sin(new_pa_gal).value
-        y_gal = r.value * np.cos(new_pa_gal).value
-        r_w = np.sqrt((xcoord-x_gal)**2 + (ycoord-y_gal)**2)
-        try:
-            p_wMi = pw_Oi(r_w, theta_prior['r_half'][icand], theta_prior)
-        except:
-            embed(header='190 of bayesian')
-
-        # Product
-        grid_p = l_w * p_wMi
-
-        # Save grids if returning
-        if return_grids:
-            grids.append(grid_p.copy())
-
-        # Average
-        p_xMis.append(np.mean(grid_p))
-
-    # Return
-    if return_grids:
-        return np.array(p_xMis), grids
-    else:
-        return np.array(p_xMis)
-
-
-def renorm_priors(raw_Oi, U):
-    """
-    Simple method to normalize the Priors
-
-    Args:
-        raw_Oi (np.ndarray):
-        U (float):
-            Prior for the FRB galaxy being undetected
-
-    Returns:
-        tuple: Normalized priors
-
-    """
-    raw_sum = np.sum(raw_Oi)
-    return (1.-U) * raw_Oi/raw_sum
-
-
-def mock_run(sky, frbs, sigR, theta, fov, scale_rhalf=2., nsigma=2.,
-             ncontam=0, seed=12345):
-    rstate = np.random.RandomState(seed) # For contaminants
-    # Coords
-    obj_coord = SkyCoord(ra=sky['ra'], dec=sky['dec'], unit='deg')
-    frb_coords = SkyCoord(ra=frbs['ra'], dec=frbs['dec'], unit='deg')
-    #
-    model_dict = {}
-    for ifrb in range(len(frb_coords)):
-        key = '{:04d}'.format(ifrb)
-        if (ifrb % 10) == 0:
-            print("Working on FRB: {}".format(key))
-        model_dict[key] = {}
-        # Grab candidates -- Assume always at least 1
-        cands = frb_coords[ifrb].separation(obj_coord) < fov
-        cand_gal = sky[cands]
-        cand_coord = obj_coord[cands]
-        # Contaminants?
-        if ncontam != 0:
-            cand_gal, cand_coord = add_contam(
-                ncontam, frb_coords[ifrb], cand_gal, cand_coord,
-                fov.to('arcsec').value, rstate=rstate)
-        # Additional
-        cand_gal['r_half'] = 1.
-        cand_sep = frb_coords[ifrb].separation(cand_coord).to('arcsec')
-        # Save
-        model_dict[key]['ncand'] = len(cand_gal)
-        model_dict[key]['cand_mag'] = cand_gal['DES_r'].data
-        model_dict[key]['theta'] = cand_sep.to('arcsec').value
-        # Raw P(Mi)
-        raw_prior_Mi = prior_Mi_n(cand_gal['DES_r'].data, cand_sep.to('arcsec').value,
-                                           cand_gal['r_half'].data, sigR.to('arcsec').value,
-                                           scale_rhalf=scale_rhalf, nsigma=nsigma)
-        # Raw P(S)
-        raw_prior_S = 1. - np.max(raw_prior_Mi)
-        # Normalize priors
-        prior_Mi, prior_S = renorm_priors(raw_prior_Mi, raw_prior_S)
-        # Calculate p(x|Mi)
-        p_xMi = px_Mi(fov.to('arcsec').value, frb_coords[ifrb],
-                               cand_coord, theta, sigR.to('arcsec').value)
-        # Calcualte p(x|S)
-        p_xS = px_Mi(fov.to('arcsec').value, frb_coords[ifrb],
-                     frb_coords[ifrb:ifrb + 1],  # Needs to be an array
-                     theta, sigR.to('arcsec').value)[0]  # Converts array to float
-        # Calculate p(x)
-        p_x = prior_S * p_xS + np.sum(prior_Mi * p_xMi)
-        # Evaluate p(Mi|x)
-        P_Mix = prior_Mi * p_xMi / p_x
-        model_dict[key]['P_Mix'] = P_Mix
-        # Evaluate p(S|x)
-        P_Sx = prior_S * p_xS / p_x
-        model_dict[key]['P_Sx'] = P_Sx
-    # Return
-    return model_dict
-
-
-def parse_model(model_dict, scale=1000, mag_lim=25.5):
-    """
-
-    Args:
-        model_dict:
-        scale:
-        mag_lim:
-
-    Returns:
-
-    """
-    model_mags = []
-    model_theta = []
-    max_PMix = []
-    #
-    for key in model_dict.keys():
-        # Observed candidates
-        for kk, p in enumerate(model_dict[key]['P_Mix']):
-            N = int(np.round(p * scale))
-            model_mags += [model_dict[key]['cand_mag'][kk]] * N
-            model_theta += [model_dict[key]['theta'][kk]] * N
-        # Unseen
-        N = int(np.round(model_dict[key]['P_Sx'] * scale))
-        model_mags += [mag_lim] * N
-        # Misc
-        max_PMix.append(np.max(model_dict[key]['P_Mix']))
-        # import pdb; pdb.set_trace()
-    # Arrays
-    model_mags = np.array(model_mags)
-    model_theta = np.array(model_theta)
-    max_PMix = np.array(max_PMix)
-    # Return
-    return model_mags, model_theta, max_PMix
